Satisfy clippy.

+++ src/packages/arithmetic.rs
@@ -304,43 +304,43 @@ mod f32_functions {
 
         #[rhai_fn(name = "+")]
         pub fn add_if(x: INT, y: f32) -> f32 {
-            (x as f32) + (y as f32)
+            (x as f32) + y
         }
         #[rhai_fn(name = "+")]
         pub fn add_fi(x: f32, y: INT) -> f32 {
-            (x as f32) + (y as f32)
+            x + (y as f32)
         }
         #[rhai_fn(name = "-")]
         pub fn subtract_if(x: INT, y: f32) -> f32 {
-            (x as f32) - (y as f32)
+            (x as f32) - y
         }
         #[rhai_fn(name = "-")]
         pub fn subtract_fi(x: f32, y: INT) -> f32 {
-            (x as f32) - (y as f32)
+            x - (y as f32)
         }
         #[rhai_fn(name = "*")]
         pub fn multiply_if(x: INT, y: f32) -> f32 {
-            (x as f32) * (y as f32)
+            (x as f32) * y
         }
         #[rhai_fn(name = "*")]
         pub fn multiply_fi(x: f32, y: INT) -> f32 {
-            (x as f32) * (y as f32)
+            x * (y as f32)
         }
         #[rhai_fn(name = "/")]
         pub fn divide_if(x: INT, y: f32) -> f32 {
-            (x as f32) / (y as f32)
+            (x as f32) / y
         }
         #[rhai_fn(name = "/")]
         pub fn divide_fi(x: f32, y: INT) -> f32 {
-            (x as f32) / (y as f32)
+            x / (y as f32)
         }
         #[rhai_fn(name = "%")]
         pub fn modulo_if(x: INT, y: f32) -> f32 {
-            (x as f32) % (y as f32)
+            (x as f32) % y
         }
         #[rhai_fn(name = "%")]
         pub fn modulo_fi(x: f32, y: INT) -> f32 {
-            (x as f32) % (y as f32)
+            x % (y as f32)
         }
     }
 
@@ -419,43 +419,43 @@ mod f64_functions {
 
         #[rhai_fn(name = "+")]
         pub fn add_if(x: INT, y: f64) -> f64 {
-            (x as f64) + (y as f64)
+            (x as f64) + y
         }
         #[rhai_fn(name = "+")]
         pub fn add_fi(x: f64, y: INT) -> f64 {
-            (x as f64) + (y as f64)
+            x + (y as f64)
         }
         #[rhai_fn(name = "-")]
         pub fn subtract_if(x: INT, y: f64) -> f64 {
-            (x as f64) - (y as f64)
+            (x as f64) - y
         }
         #[rhai_fn(name = "-")]
         pub fn subtract_fi(x: f64, y: INT) -> f64 {
-            (x as f64) - (y as f64)
+            x - (y as f64)
         }
         #[rhai_fn(name = "*")]
         pub fn multiply_if(x: INT, y: f64) -> f64 {
-            (x as f64) * (y as f64)
+            (x as f64) * y
         }
         #[rhai_fn(name = "*")]
         pub fn multiply_fi(x: f64, y: INT) -> f64 {
-            (x as f64) * (y as f64)
+            x * (y as f64)
         }
         #[rhai_fn(name = "/")]
         pub fn divide_if(x: INT, y: f64) -> f64 {
-            (x as f64) / (y as f64)
+            (x as f64) / y
         }
         #[rhai_fn(name = "/")]
         pub fn divide_fi(x: f64, y: INT) -> f64 {
-            (x as f64) / (y as f64)
+            x / (y as f64)
         }
         #[rhai_fn(name = "%")]
         pub fn modulo_if(x: INT, y: f64) -> f64 {
-            (x as f64) % (y as f64)
+            (x as f64) % y
         }
         #[rhai_fn(name = "%")]
         pub fn modulo_fi(x: f64, y: INT) -> f64 {
-            (x as f64) % (y as f64)
+            x % (y as f64)
         }
     }
